+++ RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+*:
+* Added support for *.tsv files. Default to TSV for unknown extension
+* ins-cat : By default output values of flags
+
 1.6:
 * Improved generic encoding (yaml config file, extra options)
 * Scalar variables (such as latitude,longitude,elevation)